requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
docopt==0.6.2
flatbuffers==25.1.24
fonttools==4.55.8
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.12.1
humanfriendly==10.0
idna==3.10
Js2Py==0.74
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
numpy==2.0.2
nvidia-cuda-runtime-cu12==12.8.57
onnx==1.17.0
onnxruntime-gpu==1.20.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
pillow==11.1.0
pipwin==0.5.2
protobuf==3.20.3
psutil==6.1.1
py4j==0.10.9.7
Pygments==2.19.1
pyjsparser==2.7.1
pyparsing==3.2.1
PyPrind==2.11.3
pyreadline3==3.5.4
pySmartDL==1.3.4
pyspark==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.1
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
seaborn==0.13.2
six==1.17.0
soupsieve==2.6
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
tensorrt==10.8.0.43
tensorrt-cu12==10.8.0.43
tensorrt_cu12_bindings==10.8.0.43
tensorrt_cu12_libs==10.8.0.43
termcolor==2.5.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
